+++ ydb/core/protos/grpc.proto
@@ -3,6 +3,7 @@ syntax = "proto3";
 package NKikimrClient;
 
 import "ydb/core/protos/msgbus.proto";
+import "ydb/core/protos/test_shard_control.proto";
 import "ydb/core/protos/msgbus_pq.proto";
 
 service TGRpcServer {
