+++ ydb/core/protos/msgbus.proto
@@ -661,44 +661,3 @@ message TConsoleResponse {
     optional NKikimrConsole.TStatus Status = 100;
 }
 
-message TTestShardControlRequest {
-    message TSizeInterval {
-        optional uint64 Weight = 1; // nonzero
-        optional uint32 Min = 2;
-        optional uint32 Max = 3;
-        optional bool Inline = 4; // if true, then keys are stored as inline values of KV tablet rows
-    }
-    message TTimeInterval { // Poisson distribution of events
-        optional uint64 Weight = 1; // nonzero
-        optional double Frequency = 2; // events per second
-        optional uint32 MaxIntervalMs = 3; // max interval between two events in milliseconds
-    }
-    message TCmdInitialize {
-        optional string StorageServerHost = 1; // location of storage server containing tablet actual state
-        optional int32 StorageServerPort = 2; // part of storage server location
-        optional uint64 MaxDataBytes = 3; // when total length of stored keys reaches MaxDataBytes...
-        optional uint64 MinDataBytes = 4; // then random keys are collected until total length drops below MinDataBytes
-        optional uint32 MaxInFlight = 5;
-        optional uint64 ValidateAfterBytes = 9; // number of bytes to write before starting automatic validation
-        optional uint32 MaxReadsInFlight = 10;
-        optional bool ResetWritePeriodOnFull = 11; // restart write periods when inflight is full
-        repeated TSizeInterval Sizes = 6; // distrubution of generated value size
-        repeated TTimeInterval WritePeriods = 7; // time between two events
-        repeated TTimeInterval RestartPeriods = 8; // time between automatic restarts
-        optional uint32 PatchRequestsFractionPPM = 12;
-        optional uint32 PutTraceFractionPPM = 13;
-        optional uint32 PutTraceVerbosity = 14 [default = 15];
-        optional uint32 SecondsBeforeLoadStart = 15; // number of seconds to wait before starting load
-        optional uint32 StallCounter = 16; // number of requests before issuing barrier and waiting for all of them to complete
-    }
-
-    optional uint64 TabletId = 1;
-    oneof Command {
-        TCmdInitialize Initialize = 2;
-    }
-    optional string SecurityToken = 3 [(Ydb.sensitive) = true];
-}
-
-message TTestShardControlResponse {
-    optional uint64 TabletId = 1;
-}
