+++ ydb/core/tx/schemeshard/schemeshard__operation_db_changes.h
@@ -54,6 +54,8 @@ class TStorageChanges: public TSimpleRefCount<TStorageChanges> {
 
     TDeque<TPathId> StreamingQueries;
 
+    TDeque<TPathId> TestShards;
+
     //PQ part
     TDeque<std::tuple<TPathId, TShardIdx, TTopicTabletInfo::TTopicPartitionInfo>> PersQueue;
     TDeque<std::pair<TPathId, TTopicInfo::TPtr>> PersQueueGroup;
@@ -168,6 +170,10 @@ public:
 
     void PersistStreamingQuery(const TPathId& pathId) {
         StreamingQueries.emplace_back(pathId);
+    }
+
+    void PersistTestShard(const TPathId& pathId) {
+        TestShards.emplace_back(pathId);
     }
 
     void Apply(TSchemeShard* ss, NTabletFlatExecutor::TTransactionContext &txc, const TActorContext &ctx);
