+++ ydb/core/tx/schemeshard/ut_system_names/ut_system_names.cpp
@@ -496,22 +496,20 @@ const std::vector<TCreatePathOp> CreatePathOperations({
     },
     {
         .Type = NKikimrSchemeOp::EOperationType::ESchemeOpCreateTestShard,
-        //TODO: proper check
-        .CreateRequest = nullptr,
-        // .CreateRequest = [](const TString& workingDir, const TString& path) {
-        //     auto parts = SplitPath(path);
-        //     const auto basename = parts.back();
-        //     const TString modifyScheme = Sprintf(R"(
-        //         Name: "%s"
-        //         Count: 1
-        //         StorageConfig {
-        //         }
-        //         Config: "workload: {}"
-        //     )",
-        //     basename.c_str()
-        //     );
-        //     return CreateTestShardRequest(/* txId */ 0, workingDir, modifyScheme);
-        // }
+        .CreateRequest = [](const TString& workingDir, const TString& path) {
+            const TString modifyScheme = Sprintf(R"(
+                Name: "%s"
+                Count: 1
+                StorageConfig {
+                }
+                CmdInitialize {
+                    MaxDataBytes: 1000
+                }
+            )",
+            path.c_str()
+            );
+            return CreateTestShardRequest(/* txId */ 0, workingDir, modifyScheme);
+        }
     }
 
     //NOTE: ADD NEW ENTRY ABOVE THIS LINE
